+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ Contributors to this version: Éric Dupuis (:user:`coxipi`).
 
 Bug fixes
 ^^^^^^^^^
-* Patch ``xsdba.units.convert_units_to`` with ``xclim.core.units.convert_units_to`` with `context="infer"` locally in ``xs.train`` and ``xs.adjust`` instead of using ``xc.core.settings.context``.
+* Patch ``xsdba.units.convert_units_to`` with ``xclim.core.units.convert_units_to`` with `context="infer"` locally in ``xs.train`` and ``xs.adjust`` instead of using ``xc.core.settings.context``. (:pull:`552`).
 
 
 v0.12.0 (2025-03-10)